Broke dlib-specific action out and refactored
