+++ src/chain_actions.cc
@@ -24,195 +24,12 @@
 #include <unordered_set>
 #include "utils/utils.hpp"
 
+#ifdef USE_DLIB
+#include "backends/dlib/dlib_actions.h"
+#endif
 
 namespace dd
 {
-
-#ifdef USE_DLIB
-void DlibAlignCropAction::apply(APIData &model_out,
-			     ChainData &cdata)
-    {
-      std::vector<APIData> vad = model_out.getv("predictions");
-      std::vector<cv::Mat> imgs = model_out.getobj("input").get("imgs").get<std::vector<cv::Mat>>();
-      std::vector<std::pair<int,int>> imgs_size = model_out.getobj("input").get("imgs_size").get<std::vector<std::pair<int,int>>>();
-      std::vector<std::string> cropped_imgs;
-      std::vector<std::string> bbox_ids;
-
-      // check for action parameters
-      double bratio = 0.25;
-      int chip_size = 150;
-      if (_params.has("padding_ratio")) {
-          bratio = _params.get("padding_ratio").get<double>(); // e.g. 0.055
-      }
-      if (_params.has("chip_size")) {
-          chip_size = _params.get("chip_size").get<int>(); // in pixels
-      }
-      std::vector<APIData> cvad;
-
-      bool save_crops = false;
-      if (_params.has("save_crops")) {
-          save_crops = _params.get("save_crops").get<bool>();
-      }
-
-      // iterate image batch
-      for (size_t i=0;i<vad.size();i++) {
-          std::string uri = vad.at(i).get("uri").get<std::string>();
-
-          cv::Mat cvimg = imgs.at(i);
-          dlib::matrix<dlib::rgb_pixel> img;
-          dlib::assign_image(img, dlib::cv_image<dlib::rgb_pixel>(cvimg));
-
-          int orig_cols = imgs_size.at(i).second;
-          int orig_rows = imgs_size.at(i).first;
-
-          std::vector<APIData> ad_cls = vad.at(i).getv("classes");
-          std::vector<APIData> cad_cls;
-
-          // iterate bboxes per image
-          for (size_t j=0;j<ad_cls.size();j++)
-            {
-              APIData bbox = ad_cls.at(j).getobj("bbox");
-              if (bbox.empty()) {
-                throw ActionBadParamException("align/crop action cannot find bbox object for uri " + uri);
-              }
-              APIData ad_shape = bbox.getobj("shape");
-              if (ad_shape.empty()) {
-                  throw ActionBadParamException("align/crop action cannot find shape object for uri " + uri);
-              }
-
-              // adding bbox id
-              std::string bbox_id = genid(uri,"bbox"+std::to_string(j));
-              bbox_ids.push_back(bbox_id);
-              APIData ad_cid;
-              ad_cls.at(j).add(bbox_id,ad_cid);
-              cad_cls.push_back(ad_cls.at(j));
-              const dlib::rectangle shape_rect(ad_shape.get("left").get<long>(),
-                                               ad_shape.get("top").get<long>(),
-                                               ad_shape.get("right").get<long>(),
-                                               ad_shape.get("bottom").get<long>());
-              std::vector<dlib::point> parts;
-              std::vector<double> points = ad_shape.get("points").get<std::vector<double>>();
-              for (size_t idx=0; idx < points.size()-1; idx += 2) {
-                  parts.push_back(dlib::point(static_cast<long>(points[idx]), static_cast<long>(points[idx+1])));
-              }
-
-              dlib::full_object_detection shape(shape_rect, parts);
-              dlib::matrix<dlib::rgb_pixel> r;
-              dlib::extract_image_chip(img, dlib::get_face_chip_details(shape,chip_size,bratio), r);
-              cv::Mat cropped_img = dlib::toMat(r);
-
-              // save crops if requested
-              if (save_crops)
-            {
-              std::string puri = dd_utils::split(uri,'/').back();
-              cv::imwrite("crop_" + puri + "_" + std::to_string(j) + ".png", cropped_img);
-            }
-
-              // serialize crop into string (will be auto read by read_element in imginputconn)
-              std::vector<unsigned char> cropped_img_ser;
-              bool str_encoding = cv::imencode(".png",cropped_img,cropped_img_ser);
-              if (!str_encoding)
-            throw ActionInternalException("crop encoding error for uri " + uri);
-              std::string cropped_img_str = std::string(cropped_img_ser.begin(),cropped_img_ser.end());
-              cropped_imgs.push_back(cropped_img_str);
-            }
-	  APIData ccls;
-	  ccls.add("uri",uri);
-	  if (vad.at(i).has("index_uri"))
-	    ccls.add("index_uri",vad.at(i).get("index_uri").get<std::string>());
-	  ccls.add("classes",cad_cls);
-	  cvad.push_back(ccls);
-	}
-      // store serialized crops into action output store
-      APIData action_out;
-      action_out.add("data",cropped_imgs);
-      action_out.add("cids",bbox_ids);
-      cdata.add_action_data(_action_id,action_out);
-
-      // updated model data with chain ids
-      model_out.add("predictions",cvad);
-    }
-#endif
-
-
-void ImgsCopyAction::apply(APIData &model_out,
-                               ChainData &cdata)
-    {
-
-        if (!model_out.has("input") || !model_out.getobj("input").has("imgs")) {
-            throw ActionBadParamException("copy action cannot find imgs");
-        }
-
-        APIData action_out;
-        action_out.add("data_raw_img", model_out.getobj("input").get("imgs").get<std::vector<cv::Mat>>());
-
-//        std::vector<APIData> vad = model_out.getv("predictions");
-//        std::vector<std::string> cids;
-//        for (size_t i=0; i<vad.size(); i++) {
-//            const std::string uri = vad.at(i).get("uri").get<std::string>();
-//            cids.push_back(uri);
-//            std::vector<APIData> ad_cls = vad.at(i).getv("classes");
-//            std::vector<APIData> cad_cls;
-//        }
-
-        if (model_out.getobj("input").has("cids")) {
-            action_out.add("cids", model_out.getobj("input").get("cids").get<std::vector<std::string>>());
-        }
-
-        cdata.add_action_data(_action_id,action_out);
-
-
-/*
-
-        std::vector<APIData> vad = model_out.getv("predictions");
-        std::vector<cv::Mat> imgs = model_out.getobj("input").get("imgs").get<std::vector<cv::Mat>>();
-        std::vector<std::pair<int,int>> imgs_size = model_out.getobj("input").get("imgs_size").get<std::vector<std::pair<int,int>>>();
-        std::vector<std::string> bbox_ids;
-
-        std::vector<APIData> cvad;
-
-        // iterate image batch
-        for (size_t i=0;i<vad.size();i++)
-        {
-            std::string uri = vad.at(i).get("uri").get<std::string>();
-
-            std::vector<APIData> ad_cls = vad.at(i).getv("classes");
-            std::vector<APIData> cad_cls;
-
-            // iterate bboxes per image
-            for (size_t j=0;j<ad_cls.size();j++)
-            {
-                APIData bbox = ad_cls.at(j).getobj("bbox");
-                if (bbox.empty())
-                    throw ActionBadParamException("copy action cannot find bbox object for uri " + uri);
-
-                // adding bbox id
-                std::string bbox_id = genid(uri,"bbox"+std::to_string(j));
-                bbox_ids.push_back(bbox_id);
-                APIData ad_cid;
-                ad_cls.at(j).add(bbox_id,ad_cid);
-                cad_cls.push_back(ad_cls.at(j));
-            }
-            APIData ccls;
-            ccls.add("uri",uri);
-            if (vad.at(i).has("index_uri"))
-                ccls.add("index_uri",vad.at(i).get("index_uri").get<std::string>());
-            ccls.add("classes",cad_cls);
-            cvad.push_back(ccls);
-        }
-        // store serialized crops into action output store
-        APIData action_out;
-        action_out.add("data_raw_img",imgs);
-        action_out.add("cids",bbox_ids);
-        cdata.add_action_data(_action_id,action_out);
-
-        // updated model data with chain ids
-        model_out.add("predictions",cvad);
-*/
-    }
-
-
-
   void ImgsCropAction::apply(APIData &model_out,
 			     ChainData &cdata)
     {
@@ -342,6 +159,32 @@ void ImgsCopyAction::apply(APIData &model_out,
     
     // updated model data
     model_out.add("predictions",cvad);
+  }
+
+  void ChainActionFactory::apply_action(const std::string &action_type,
+                          APIData &model_out,
+                          ChainData &cdata) {
+    std::string action_id;
+    if (_adc.has("id"))
+        action_id = _adc.get("id").get<std::string>();
+    else action_id = std::to_string(cdata._action_data.size());
+    if (action_type == "crop") {
+        ImgsCropAction act(_adc, action_id, action_type);
+        act.apply(model_out, cdata);
+    } else if (action_type == "filter") {
+        ClassFilter act(_adc, action_id, action_type);
+        act.apply(model_out, cdata);
+    }
+#ifdef USE_DLIB
+        else if (action_type == "dlib_align_crop")
+{
+DlibAlignCropAction act(_adc,action_id,action_type);
+act.apply(model_out,cdata);
+}
+#endif
+    else {
+        throw ActionBadParamException("unknown action " + action_type);
+    }
   }
 
 } // end of namespace
